+++ test/foundry/strategies/DonationVotingMerkleDistributionBase.t.sol
@@ -446,7 +446,9 @@ contract DonationVotingMerkleDistributionBaseMockTest is
     function testRevert_updatePoolTimestamps_INVALID() public {
         vm.expectRevert(INVALID.selector);
         vm.prank(pool_admin());
-        strategy.updatePoolTimestamps(block.timestamp - 1, registrationEndTime, allocationStartTime, allocationEndTime);
+        strategy.updatePoolTimestamps(
+            uint64(block.timestamp - 1), registrationEndTime, allocationStartTime, allocationEndTime
+        );
     }
 
     function testRevert_withdraw_NOT_ALLOWED_30days() public {
